[BUGFIX:BP:11.5] Do not index missing fields
In case the data used for a Solr field is not present
in the data (e.g. if it is conditionally set via a
custom indexer), the field will no longer be created.

Previously the data was accessed anyways, resulting in
a PHP warning which would abort the indexing in PHP 8.

Now a check has been implemented to verify that the accessed
variable actually exists in configuration and data.

Fixes: #3366