!!![TASK] Introduce RootPageRecordNotFoundException
Introduces a specific exception for missing root page records.

If the page record for a root page check in RootPageResolver couldn't
be found a generic InvalidArgumentException was thrown, this is now
replaced with the specific RootPageRecordNotFoundException.

This is a breaking change as extensions might expect the
InvalidArgumentException and try to handle issues on that basis, this
extension have to be updated and consider the new exception.

Resolves: #3500