[TASK] Handle Solr connection exception
Solr requests may fail due to server errors or faulty queries,
thrown exceptions or just an invald status code indicate such
errors.

This commit ensures that exceptions are handled and a valid
response with a meaningful status code is returned. Handling
is centralised in SolrService so that other components just
have to deal with the status code.

Additionally the GarbageCollection is considering this status
and is logging missing index update.

Resolves: #2903