!!![TASK] Drop custom translate view helper
Drops custom TranslateViewHelper as this just provides named
placeholders in labels and with TYPO3 12 we had to completly
reimplement the translate feature as the TYPO3 core version
cannot be used as base anymore.

This is a breaking change for all users that are using this ViewHelper
and own labels and markers like @resultsFrom or @resultsTo. Please
replace this markers with default markers like "%s" or "%1$s" and use
the default TYPO3 translate view helper.

Relates: #3376