!!![TASK:T12] Don't initialize TSFE in TYPO3 BE (Record Monitoring) context
EXT:solr does not want really the initialized TSFE object on multiple places at all. The TSFE object was necessary in EXT:solr versions, which used the TypoScript conditions for configure/map the Apache Solr cores to appropriated TYPO3 languages.  The indexing stack required this approach as well to be able to index  different pages and/or its records in single PHP-Process.

Nova-days this all hacks not required on TYPO3 12+ LTS anymore,  so the whole TSFE-Fake initialization stack will be removed and  "isolated TYPO3 Application" approach will be used. This will lead to more isolated and less exceptional behavior on TYPO3 components on EXT:solr side/feature set.

Relates: #3376