[!!!][TASK] Remove deprecated functionality
This change removes deprecated features of EXT:solr

a) PageDocumentPostProcessor

The interface PageDocumentPostProcessor
is removed, along with the registration
in ext_localconf.php.

The PostProcessor is now removed in favor
of the PageIndexerDocumentsModifier interface,
and can be used already in EXT:solr v11.

Check in your extensions' ext_localconf.php
and look for
`$GLOBALS['TYPO3_CONF_VARS']['EXTCONF']['solr']['Indexer']['indexPagePostProcessPageDocument']`
to see if you are affected.

In addition, some rather unintrusive, deprecated
methods in UrlHelper have been removed
as UrlHelper is just a small
extension for UriInterface, and might
vanish altogether in the future.

Relates: #3376