[BUGFIX] Sanitize frequent and last searches output
Wraps output of frequent searches and last searches in
format.htmlentities, to prevent XSS. In rare scenarios
there might be issues with unsanitized output of frequent
searches and/or last searches, this issues are now solved.

The StatisticsWriterProcessor is already sanitizing the stored
terms, but as it's possible to use any custom table it should be
ensured the data is sanitized.

If last searches is activated and configured for global mode,
there might be issues too. Though it's required that the prepared
query delivers result documents, as EXT:solr will only store the
terms if results are found, the output should be sanitized too.

Resolves: #3589