[TASK] Integrate phpstan
This change adds phpstan for static code analysis.

The current code base (especially the tests),
define wrong phpdoc comments, which is the main work
to achieve in this patch.

In order to make tests / mocking behaviour easier,
the phpunit extension is also installed.

For this reason, the "getDumbMock" need to be
removed for "createMock" from phpunit.

The test can be executed like this:

* composer install --prefer-source
* composer tests:phpstan