[FEATURE] Add new Events for Indexing
This change adds new PSR-14 events in favor of:

* $GLOBALS['TYPO3_CONF_VARS']['EXTCONF']['solr']['IndexQueueIndexer']['indexItemAddDocuments']
* $GLOBALS['TYPO3_CONF_VARS']['EXTCONF']['solr']['Indexer']['indexPageAddDocuments']
* $GLOBALS['TYPO3_CONF_VARS']['EXTCONF']['solr']['IndexQueueIndexer']['preAddModifyDocuments']

This way, EXT:solr migrates to PSR-14 events and adds a more
defined way to utilize the PHP API and to help extension authors
to understand what is used in EXT:solr.

This still needs to be defined, whether we want to remove
the hooks in this patch as well.

Also, documentation needs to be adapted.