[!!!][FEATURE] Add new PSR-14 instead of SubstitutePageIndexer
$GLOBALS['TYPO3_CONF_VARS']['EXTCONF']['solr']['Indexer']['indexPageSubstitutePageDocument']
is removed in favor of a new PSR-14 event:

"ApacheSolrForTypo3\Solr\Event\Indexing\AfterPageDocumentIsCreatedForIndexingEvent"

For this reason, the SubstitutePageIndexer Interface is removed.

At the same time, the "Typo3PageIndexer" class is now merged
into FrontendHelper/PageIndexer to avoid cross-cutting conerns
between classes and to show the dependencies to TSFE properly.