[TASK] Use Apache Solr 9.2 for EXT:solr 12.0
Updates to Apache Solr 9.2, this includes the required
configuration updates. luceneMatchVersion is sill set to 8.5.0
to prevent changes in search behaviour.

To avoid security issues described in CVE-2022-39135 the SQL
module is not enabled, you have to enabled the module manually
if required.

Note: With Apache Solr 9 the following components are no longer
available and you have to adapt the configuration if needed. No
longer available components are:

1) Data Import Handler (DIH)
DIH is an independent project now; it is no longer a part of Solr

2) VelocityResponseWriter
VelocityResponseWriter is an independent project now; it is no
longer a part of Solr. This encompasses all previously included
/browse and wt=velocity examples.

This update is considered as breaking, as a new Solr server
and configset is required, but though not recommended you
can still use your old configset and Solr 8.5 if an update is
not possible.

Fixes: #3661