[!!!][TASK] Rework frontend indexing helpers
This change cleans up a few parts within the FrontendHelper
logic when indexing a page in the frontend:

- The PageIndexerInitialization and PageIndexerFinisher middleware
  are now merged.

- The PageIndexerRequestHandler contains no state

- The PageIndexerRequest is added to the PSR-7 Request as attribute
  "solr.pageIndexingInstructions" so it
  can be retrieved in various places.

- The AbstractFrontendHelper is removed, as the FrontendHelper
  middleware is thinned out. Activation and Deactivation rather
  works as initialization and flushing

This allows us to use DI in various places we hadn't had a chance
before.

In addition, the AuthorizationService getGroupsFE is removed
as it has been already replaced by a PSR-14 event.