[BUGFIX] fix null-pointer-exception in AbstractStrategy
Our indexing-queue contained an item for a hidden root page, that hasn't
been localized yet. When updating the hidden root page, the determined
site for the index queue item
in AbstractStrategy::deleteIndexDocuments() is null, which triggers
a PHP-Runtime error when getSolrConfiguration() is called on it.

We weren't able to figure out how this item ended up in the queue,
and we also weren't able to add it manually using the
"Index Queue" TYPO3 backend module.

Our instance was a TYPO3 11 with solr 11.5.3.