[BUGFIX:BP:11.5] Fix EXT:solr route enhancer
SolrRoutingMiddleware was limited to indexing processs, but
as the SolrFacetMaskAndCombineEnhancer requires this middleware
to add the required parameters from the route path, this limitation
causes that the route enhancer is ineffective.

This commit fixes this issue by reenabling the middleware for
frontend requests and also ensures that route paths can be resolved
correctly if PageTypeSuffix enhancer is active.

Ports: #3742
Relates: #3202
Resolves: #3741