[TASK:BP:11.6] Ensure recursive page update on page movement
If a page is moved, subpages have to be considered, e.g. as rootline
and slug might have changed. Also the MountPagesUpdater has to be
triggered to ensure mounted pages will be updated too.

This commit adds the missing recursive updates, an extension to the
RecordMovedEvent and PageMovedEvent is done allowing that recursive
updates on move operations will only be done if the parent page of
the page has changed.

Ports: #3787
Resolves: #206