[TASK] Require TYPO3 12.4.3 to get typo3fluid/fluid >= 2.9.2
Due to an internal bug in typo3fluid/fluid the pagination
might fail, this bug is fixed in typo3fluid/fluid 2.9.2.

We now require TYPO3 12.4.3 to get typo3fluid/fluid >= 2.9.2,
so it is ensured the right version is used and the pagination
works as expected.

The failing test is reenabled.

Resolves: #3691