!!![CLEANUP] Remove obsolete TYPO3 registry usages
TYPO3 registry was mainly used for connection storage, which is
obsolete since the introduction of the Site handling. Luke limit
read by the InfoModuleController was never set in the registry,
so for now the fallback value of 20000 is used.

This is a breaking change as the SiteRepository expected the
registry to be passed, which is no longer necessary.

Resolves: #3792