!!![TASK] Clean and optimize frontend helper: PageIndexer
Cleans and optimizes the frontend helper "PageIndexer" integration
test.

Failure test for non existing pages/sites is refactored and
PageIndexer adapted, even if AfterCacheableContentIsGeneratedEvent
is not dispatched e.g. as the page is not found, the response will
always contain the pageIndexed flag.

Additionally the methods of PageIndexer that were defined as public
to simplify testing are now protected, the tests are adapted
accordingly. If you have used this methods, you have to adapt your
code. Affected methods:
- setupConfiguration
- index
- indexPage

Relates: #3735