!!![FEATURE] Allow using complex filter values in FlexForm
With this change it is now possible to use complex query in FlexForm filter value.

The old behaviour is still working,
so if a string value contains space(s) and no special characters of the solr query parser,
the string is always wrapped with double quotes.
But if the string contains special characters no wrapping happen
special characters are: `+ - && || ! ( ) { } [ ] ^ " ~ * ? : \`.

There is some cases where this change can break,
for example if the filter value is something like
`toto AND tata` or `music (rock)` or `my "flow" is`.
Here the wrapping and the escaping of the inner double quote have to be manually updated like this
`"toto AND tata"` or `"music (rock)"` and `"my \"flow\" is"`.

Fixes: #2151