!!![BUGFIX] Exception with tx_solr_statistics after latest TYPO3 security update
**Note:** This change requires the database schema update.

With the Security update at November, 14th 2023 the cookie generation changed in the TYPO3 core.

This leads to an exception, when executing the search and statistics are enabled.

Reason is, that the generated cookies are longer than the old ones and do not fit in the tx_solr_statistics.cookie any more. The available length is 32 chars, TYPO3 core allows up to 190.

This change increases the length of `tx_solr_statistics.cookie` field to 255 characters. Beside of that, the removed in TYPO3 12 method to fetch the fe_user cookie value is adjusted to new TYPO3 API and is functional now.

Fixes: https://github.com/TYPO3-Solr/ext-solr/issue/3877
Ports: https://github.com/TYPO3-Solr/ext-solr/pull/3880
Ports: #13