[TASK] Adjust integration test environment
StoragePermissionsAspect triggered on TSFE initialization if an
absRefPrefix is set, expects a BE_USER object and will fail if
not present. Failure only occurs if a prefix is set or automatically
determines as TYPO3_SITE_URL may vary, to prevent issues a BE_USERS
is initialized in the affected tests.

Relates: #15