[BUGFIX] follow-up: false as field value is not allowed
This change is a follow-up of #3901, which was introduced to allow following values in Apache Solr fields:
* numeric 0
* boolean false
which lead to countless-unwanted numeric fields with 0 value in Apache Solr documents.
The background: TypoScript can return the strings only, and all empty values lead to numeric 0.

This change tries to avoid mispelled casting of numbers and lets the 
`AbstractIndexer::ensureFieldValueType()` return PHPs `null` if the value isn't a number. 

Relates: #3900, #3901