[BUGFIX] numeric facet range slider sends lot of requests to server
Previously the numeric range slider updated the UI and the facet setting continously during sliding. This caused a lot of get requests to the server. This PR changes the desired behavior to send update requests only once the user releases the slider. The UI still gets updated continuously by using the separate slide-event for that.

Fixes: #4082