[BUGFIX] Ensure index document is deleted
If an index document has no representation in the index queue,
EXT:solr fails to remove the item on deletions or updates of
the corresponding record in TYPO3.

By adjusting the checks in the record monitor, garbage remover
and update handler this issue is resolved. The
ConfigurationAwareRecordService is no longer using
BackendUtility::getRecord() as this would return hidden records.

Resolves: #4107