[FIX] access restrictions stack for TYPO3 13
This change fixes the access restrictions stack within TYPO3 13.
There were multiple troubles in classes and in integration tests.
The page indexer stack on live system runs in multiple isolated processes:

1. CLI
2. first FE request to get FE-groups
3. multiple FE requests to get real content from page for desired FE-groups combination

unlike on integration test running the page indexer steps allways on single process. 
This requires to unset the produced state on each step of indexing process.
The runtime cache was interfering between the steps 
and did things that had nothing to do withhin the own step.

---

The main problem within access restriction stack on page indexing 
was the registration of Event-Listeners methods within `UserGroupDetector` with `#[AsEventListener]` on EXT:solr. 
Due of current troubles with circular dependencies 
the auto-wiring and auto-configuration was disabled in Servvices.yaml globally. 
It requires to enable this features manually for classes, which was done for `UserGroupDetector` class.

Relates: #3995
Fixes: #4161