[TASK] Adjust mount point indexing
Mount point indexing and corresponding tests have been adjusted for
TYPO3 13. Mount points are supported in general and the mounted pages
will be indexed like standard pages.

But there is a point to consider: Mounted pages from a pagetree
without a site configuration cannot be indexed, in fact TYPO3
currently can't mount a page from a page tree without a site
configuration and an exeception occurs. The behavior is intentionally
designed this way in TYPO3 core, the background is that it is not
possible to specify the languages of the mounted page tree without
Site Configuration.

Test procedure in PageIndexerTest has been improved and now uses
the PageUriBuilder to build the url to index, instead of building
the url in the test itself.

Note: We require at least TYPO3 13.4.2, as this version contains some
bugfixes that address problems with the determination of TypoScript
and the site configuration of mounted pages.

Resolves: #4160