[TASK] Adjust mount point indexing
Mount point indexing and corresponding tests have been adjusted for
TYPO3 13. Mount points are supported in general and the mounted pages
will be indexed like standard pages.

But there is a point to consider: Mounted pages from a pagetree
without a site configuration cannot be indexed, in fact TYPO3
currently can't mount a page from a page tree without a site
configuration and an exeception occurs.

Test procedure in PageIndexerTest has been improved and now uses
the PageUriBuilder to build the url to index, instead of building
the url in the test itself.

Resolves: #4160