[RELEASE] 11.5.7
This is a maintenance release for TYPO3 11.5 and the last release
that supports Apache Solr 8.11.

Next EXT:solr release for TYPO3 11.5 will be 11.6.0, it will contain
support for Apache Solr 9 and some breaking improvements.

This maintenance release for TYPO3 11.5 LTS contains:

- [TASK] Modify column changed to int(11) unsigned, in respect…
- [TASK] Full fixed french translation
- [TASK] Fix tests on 11.5.x @2024.05.07
- [BUGFIX:P:11.5] Cast pageType and suggestPageType to int in
  SearchFormViewHelper
- [BUGFIX] Typos in last FR translation PR
- [BUGFIX:P:11.5] Fix range string calculation in DateRange facet
- [BUGFIX:11.5] scheduler task "Optimize index of a site" is not
  functional
- [TASK] Apache Solr 8.11.4 compatibility

Please read the release notes:
https://github.com/TYPO3-Solr/ext-solr/releases/tag/11.5.7

---

How to Get Involved

There are many ways to get involved with Apache Solr for TYPO3:

* Submit bug reports and feature requests on GitHub
* Ask or help or answer questions in our Slack channel
* Provide patches through pull requests or review and comment on
existing pull requests
* Go to www.typo3-solr.com or call dkd to sponsor the ongoing
development of Apache Solr for TYPO3

Support us by becoming an EB partner:
https://shop.dkd.de/Produkte/Apache-Solr-fuer-TYPO3/

or call:
+49 (0)69 - 2475218 0