!!![FIX] Docker execution order issue for as-sudo tweaks
This PR renames the file `as-sudo-tweaks.sh` to `0_as-sudo-tweaks.sh`
and moves the folder
`/docker-entrypoint-initdb.d/as-sudo/`
to
`/docker-entrypoint-initdb.d-as-sudo/` to fix the execution order issue
when setting the correct file permissions when starting the docker container,
leading to a `Operation not permitted` errors.

More details see:
* https://github.com/TYPO3-Solr/ext-solr/issues/3837#issuecomment-2461668377.
* https://github.com/TYPO3-Solr/ext-solr/pull/4219#issuecomment-2622600937

This change requires adjustments in your Docker setup, only if you modified:
files `/docker-entrypoint-initdb.d/as-sudo/` in folder
or `/docker-entrypoint-initdb.d/as-sudo-tweaks.sh`.

Make sure to use:
* `/docker-entrypoint-initdb.d/0_as-sudo-tweaks.sh` instead of `/docker-entrypoint-initdb.d/as-sudo-tweaks.sh`
* `/docker-entrypoint-initdb.d-as-sudo` instead of `/docker-entrypoint-initdb.d/as-sudo/`

Fixes: #3837
Ports: #4219
Ports: https://github.com/TYPO3-Solr/ext-solr/pull/4307