[BUGFIX] Fix record monitoring if site is missing
Updates of records or pages lead to an SiteNotFoundException if
no site is configured. This is caused by the GarbageHandler and/or
DataUpdateHandler.

This issue is fixed by adding additional checks, skipping the
monitoring is save as change are not relevant for solr if no site
is present, this includes mount pages as TYPO3 cannot mount pages
outside of a configured site.

Relates: 4160