[DOCS] workaround for version matrix
Currently it is not possible to use spans in tables, like it was possible on Sphix-Variant.
See: https://github.com/phpDocumentor/guides/issues/1187

On the test `test_documentation` CI job, we want to generate the docs twice to be able to recognize the errors and warnings in the docs.
See: https://github.com/phpDocumentor/guides/issues/1188

Fixes: #4204