[FEATURE] Add an event for modifying the domain used for a site
When running the TYPO3 backend on a different domain the the TYPO3 frontend via a hostVariant it is impossible to get the correct domain in the backend. This has an impact on indexing, clearing the index, and much more. This event allows to modify the domain used for a site.