[BUGFIX] Fix mount point garbage collection
EXT:solr failed to delete indexed mounted pages if a mount
point is deleted or deactivated, this issue is fixed by extending
the PageStrategy.

Additionally it's ensured, that changes to the slug of a mount page
triggers a reindexing of the mounted pages as the indexed slug
has become invalid.

Ports: #4315
Resolves: #4314