[BUGFIX] Ensure rootline order is respected detecting closest template
When detecting the closest template inside a rootline, it is necessary,
respecting the order of the rootline itself.

The current database call has a random sorting, which does not respect
the ordering, therefore, could return a wrong dataset if multiple
templates are given in the root line.

To ensure the order is respected, every page inside rootline is called
single against the database and returns the pid if a template record
exists.