[TASK] add test case for hidden translation in content-fallback and strict mode
This change fixes the issue on integration-tests stack.
Previously the integration-tests behave different from real TYPO3 application.
Multiple tests, where a TYPO3 components with usage of Core-Context did not fail properly,
because the core-context was always initialized for a frontend application,
but EXT:solr in indexing context always uses Command or Backend applications.
So in real TYPO3 applications the `visibility` aspect was always set to show hidden/disabled records,
but on test-stack not...

It adds the missing test case for hidden translations in modies:

### In content-fallback mode:
1. The translated and not hidden record must be shown in FE and Search
2. The translated but hidden record must be shown as replaced by original in FE and Search
3. The record without translation must be shown and replaced by original in FE and Search

### In strict mode:
1. The translated and not hidden record must be shown in FE and Search
2. The translated but hidden record must NOT be shown in FE and Search
3. The record without translation must NOT be shown in FE and Search

Relates: #4427, #4409