[BUGFIX] Add setting to disabled facets
There is an issue resolving the typoscript configuration properly. Currently there is no way to override the default tx_solr configuration in the context of another content element/plugin.
The problem occurs in the AbstractBaseController initializeAction. If e.g. __UNSET is used to unset a facet this is properly done in line 169. As intended, the facet is removed, but also the __UNSET rule.
In line 179 mergeSolrConfiguration() the facet is added again from the "default" plugin configuration as the __UNSET rule is now removed. This might affect other settings than facets too (e.g. sorting.options).

Resolves: #4419 #2333