[BUGFIX] Describe array shape of findTranslationOverlaysByPageId correctly
The previously existing shape means: "an array with exactly one item". However it should be "an array with an arbitrary number of records".

We have extended this method in our code base and try to return `[]` for some cases however phpstan thinks that does not comply with the parent method's array shape.